Beginning of the actual program
I got started on actually drawing diagrams for the given schlafli symbol
but I have a few bugs where only a few lines show up and at the wrong
angles. I already checked and I am confident that the locations of the
vertices are correct but I still can't find what is causing the issue.